# Speed up IndexedDISI Sparse #AdvanceExactWithinBlock for tiny step advance