--- a/lucene/CHANGES.txt
+++ b/lucene/CHANGES.txt
@@ -131,7 +131,7 @@ Improvements
 Optimizations
 ---------------------
 
-* GITHUB#12324: Speed up Sparse block advanceExact with tiny step in IndexedDISI. (Guo Feng)
+* GITHUB#12324: Speed up sparse block advanceExact with tiny step in IndexedDISI. (Guo Feng)
 
 * GITHUB#12270 Don't generate stacktrace in CollectionTerminatedException. (Armin Braun)